Migration to ImageService

+++ org.jcryptool.analysis.graphtools/src/org/jcryptool/analysis/graphtools/Bar.java
@@ -225,7 +225,7 @@ public class Bar {
 	public void drawBar(Rectangle myBox, double myHeight, double myWidth, MColor myColorMainbar, GC myGC) {
 		int barWidth = (int) Math.round(myWidth * myBox.width);
 		int barHeight = (int) Math.round(myHeight * myBox.height);
-		int topleftX = (int) Math.round((double) (myBox.width - myWidth * myBox.width) / 2 + myBox.x);
+		int topleftX = (int) Math.round((myBox.width - myWidth * myBox.width) / 2 + myBox.x);
 		int topleftY = myBox.y + myBox.height - barHeight;
 		Rectangle myBar = new Rectangle(topleftX, topleftY, barWidth, barHeight);
